chore: add support for the openebs plugin
the openebs plugin supercedes the mayastor
plugin.
However it is not a dropin replacement.
It is functional for all openebs IO engines,
and stem commands needs to further qualify
which IO engine the command is for.

At present invocations of the plugin in the
code base only exist for mayastor.

Refactor by adding functions
 -  GetMayastorPluginCmd returns and exec cmds with the
    correct argument list
 - GetMayastorPluginCmdString - needed for invocation of
   support shell scripts using the plugin
For the argument list is determined by name.
The assumption is that if the plugin name is
not kubectl-mayastor then it is the new plugin.

Refactor to replace most uses of GetPluginPath,
the exceptions being
  - upgrade
  - plugin version check commands

Also add support to create a config map for mayastor
diskpools.

Compatibilty with kubectl-mayastor is maintained

Signed-off-by: Blaise Dias <blaise.dias@datacore.com>